# openapi.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: Delivery de comida backend API.
  contact:
    email: car.bis@hotmail.com
  version: 1.0.0
...
servers:
  - url: https://localhost:3000
  - url: http://localhost:3000
...
paths:
  /users:
    post:
      tags:
        - User
      summary: Create user
      description: Creates a new User in the DataBase
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      security: []
      responses:
        200:
          description: User correctly added to database
        400:
          description: Error validating input data
        409:
          description: Username already exists, please pick another
        default:
          $ref: "#/components/responses/Default"

    get:
      tags:
        - User
      summary: Gets list of users(Only Admin) or User account details (Not Admin)
      description: Gets a list of all the existing users in the DataBase if the user
        sending the request is an admin, otherwise the response only targets the active
        user.
      operationId: getUserInfo
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    put:
      tags:
        - User
      summary: Edits active user data
      operationId: editActiveUserData
      requestBody:
        description: Fields to change
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/User"
      responses:
        200:
          description: Successful operation
        400:
          description: Invalid username/password supplied
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          description: Username or email already registered
        default:
          $ref: "#/components/responses/Default"

    delete:
      tags:
        - User
      summary: Disables user account (validation via token)
      operationId: deleteActiveUser
      responses:
        200:
          description: User account disabled
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
  /users/login:
    get:
      tags:
        - User
      security: []
      summary: Logs user into the system
      operationId: loginUser
      parameters:
        - name: username or email
          in: query
          description: The user name or email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        400:
          description: Invalid username/password supplied
        401:
          $ref: "#/components/responses/Disabled"
        default:
          $ref: "#/components/responses/Default"
  /users/:username:
    get:
      tags:
        - User
      summary: Get user by user name (only Admin)
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    put:
      tags:
        - User
      summary: Edit by username (only Admin)
      description: Edits a username fetched by its username (Admin Only). The full object can be sent or only the parameters needing to be modified.
      operationId: updateUserAdmin
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Updated user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        200:
          description: "User {username} was modified correctly"
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    delete:
      tags:
        - User
      summary: Delete(invalidate) user (only Admin)
      description: This can only be done by a logged in admin.
      operationId: deleteUserAdmin
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        200:
          description: "User {username} was disabled correctly"
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"
  /products:
    get:
      tags:
        - Products
      summary: Get list of all available products
      description: Returns a list of products (does NOT list disabled products)
      operationId: getProducts
      parameters:
        - name: search
          in: query
          description: Keywords to search for
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    post:
      tags:
        - Products
      summary: Adds a product (only Admin)
      description: Adds a product to database
      operationId: addProduct
      requestBody:
        description: Keywords to search for
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        200:
          description: Product added succesfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  example: -2
        400:
          description: Error validating input data
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
  /products/:id:
    get:
      tags:
        - Products
      summary: Gets product by Id
      description: Returns a specific product detail, fetched by it's ID
      parameters:
        - name: id
          in: path
          description: The ID of the product that needs to be fetched.
          required: true
          schema:
            type: string
      operationId: getProductById
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        400:
          description: Error validating input data
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    put:
      tags:
        - Products
      summary: Edits product by Id (only Admin)
      description: Edits a specific product fetching it by it's ID
      parameters:
        - name: id
          in: path
          description: The ID of the product that needs to be fetched.
          required: true
          schema:
            type: string
      operationId: editProduct
      requestBody:
        description: Modification data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        200:
          description: "Product with id {productId} modified correctly"
        400:
          description: Error validating input data
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    delete:
      tags:
        - Products
      summary: Disables product by Id (only Admin)
      description: Disables a specific product, fetched by it's ID
      parameters:
        - name: id
          in: path
          description: The ID of the product that needs to be fetched.
          required: true
          schema:
            type: string
      operationId: deleteProduct
      responses:
        200:
          description: "Product with id ${productId} was disabled correctly"
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"
  /orders:
    get:
      tags:
        - Orders
      summary: Get list of orders
      description: Returns a list of products. If user is admin returns a list of all orders, if user is not admin returns a list of their orders only.
      operationId: getOrders
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    post:
      tags:
        - Orders
      summary: Adds a product
      description: Adds a product to database
      operationId: addOrder
      requestBody:
        description: Order to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOrder"
        required: true
      responses:
        200:
          description: Order modified succesfully
        400:
          description: Error validating input data
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 100
        username:
          type: string
          example: sabrina
        password:
          type: string
          example: sabri123
        email:
          type: string
          example: sabrina@gmail.com
        delivery_address:
          type: string
          example: calle 123
        full_name:
          type: string
          example: sabrina rojas
        phone:
          type: integer
          example: 4585756
    UserInfo:
      type: array
      items:
        type: object
        properties:
          user_id:
            type: integer
            example: 1
          username:
            type: string
            example: carla123
          full_name:
            type: string
            example: carla conte
          email:
            type: string
            example: carla@gmail.com
          phone:
            type: integer
            example: 4252525
          delivery_address:
            type: string
            example: calle 18
          is_admin:
            type: boolean
            example: 0
          is_disabled:
            type: boolean
            example: 0
...
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"
...
    Orders:
      type: array
      items:
        $ref: "#/components/schemas/Order"
...
    Token:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
...
    Send:
      type: string
      example: "Error message"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reused Responses
  responses:
...
    Disabled:
      description: Invalid request, user account is disabled
...
    NotFound:
      description: The specified resource was not found
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
    UnauthorizedAdmin:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
            Not Admin:
              summary: Unauthorized - Not an admin
              value: Unauthorized - Not an admin
security:
  - bearerAuth: []